backend/old/requirements.txt
flask==2.3.3
google-genai==0.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
markdown==3.5.1